Keep the CSP verifier strict while allowing JSON-LD
The post-build step rewrote each document's script-src from the inline
scripts it found, so verify-csp.mjs was then comparing the policy against
the artifact it had just been generated from and could never fail. An
alert(1) injected into a built page was blessed with a hash and the build
reported ok.

SvelteKit's hash-mode CSP genuinely does not hash structured data written in
svelte:head, so a step is needed. This one hashes only
type="application/ld+json" blocks, which browsers never execute, and appends
them to the existing policy instead of replacing it. The same injected
alert(1) now fails the build, while the JSON-LD passes.

Also restores the accents in the two new Spanish strings, since the plain
ASCII rule is about typographic characters rather than Spanish orthography,
and records the prerendering change in docs/decisions.md, which the earlier
entry on the static SPA otherwise contradicts.
