+++ frontend/package.json
@@ -8,7 +8,7 @@
 	},
 	"scripts": {
 		"dev": "vite dev",
-		"build": "vite build && node scripts/sync-csp-hashes.mjs && node scripts/verify-csp.mjs",
+		"build": "vite build && node scripts/add-jsonld-csp-hashes.mjs && node scripts/verify-csp.mjs",
 		"preview": "vite preview",
 		"prepare": "svelte-kit sync || echo ''",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
